Revert "Merge branch 'CC-17-Making-the-main-menu-options-functional' into CC-5-Selecting-An-Option-On-The-Main-Menu-Should-Be-Functional"
This reverts commit bc81d3d884d98fe45124edcc828ad5bdef3ae1ff, reversing
changes made to aba8f1c222203c6b458e776eba6fe71878ffbbfd.
