+++ Gemfile
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 source "https://rubygems.org"
 
 gem "sinatra-activerecord"
@@ -12,3 +13,4 @@ gem 'uri'
 gem 'net-http'
 gem 'json'
 gem 'colorize'
+
