+++ config/environment.rb
@@ -1,6 +1,5 @@
 require 'bundler'
-
 Bundler.require
 
-# ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'db/development.db')
-# require_all 'lib'
+#ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'db/development.db')
+#require_all 'lib'
