--- lib/models/Compare_Crypto.rb
@@ -1,7 +1,0 @@
-class CompareCrypto
-
-    def welcome_user
-        puts "Welcome to Compare Crypto!"
-    end
-
-end
