--- lib/models/My_Portfolio.rb
@@ -1,7 +1,0 @@
-class MyPortfolio
-
-    def welcome_user
-        puts "Welcome to My Portfolio!"
-    end
-
-end
