--- lib/models/Price_Check.rb
@@ -1,7 +1,0 @@
-
-class PriceCheck
-    def welcome_user
-        puts "Welcome to Price Check!"
-    end
-
-end
