--- spec/run_spec.rb
@@ -1,3 +1,0 @@
-require_relative '../lib/run.rb'
-require_relative './spec_helper.rb'
-require_relative '../config/environment'
